Move django migrations test plugin to own module
Migrations are now enabled by default.
